backend/requirements.txt
# FastAPI 核心
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.12

# Word 文档处理
python-docx==1.1.2
docxtpl==0.16.7

# 数据验证
pydantic==2.9.2
pydantic-settings==2.5.2

# 数据库
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
alembic==1.13.3

# 认证
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# 工具
python-dotenv==1.0.1
httpx==0.27.2
tenacity==9.0.0

# 存储
minio==7.2.9

# 缓存
redis==5.2.0

# 其他
aiofiles==24.1.0
